Crate hdf5

Re-exports

pub use hdf5_derive::H5Type;

Modules

dataset

file

plist

types

Macros

h5lock
Run a potentially unsafe expression in a closure synchronized by a global reentrant mutex.

h5call
Convert result of HDF5 call to Result; execution is guarded by a global reentrant mutex.

Structs

Container

Dataset
Represents the HDF5 dataset object.

DatasetBuilder

Dataspace
Represents the HDF5 dataspace object.

Datatype
Represents the HDF5 datatype object.

File
HDF5 file object.

FileBuilder
File builder allowing to customize file access/creation property lists.

Filters
HDF5 filters and compression options.

Group
Represents the HDF5 group object.

Location
Named location (file, group, dataset, named datatype).

Object
Any HDF5 object that can be referenced through an identifier.

PropertyList
Represents the HDF5 property list.

Reader

Writer

Enums

Conversion

Error
The error type for HDF5-related functions.

Traits

Dimension
A trait for the shape and index types.

H5Type

Functions

from_id⚠

is_library_threadsafe
Returns true if the HDF5 library is threadsafe.

library_version
Returns the runtime version of the HDF5 library.

silence_errors

Type Definitions

Ix
A scalar integer type used by Dimension trait for indexing.

Result
A type for results generated by HDF5-related functions where the Err type is set to hdf5::Error.