Struct hdf5::dataset::DatasetBuilder

pub struct DatasetBuilder<T> { /* fields omitted */ }

Methods

impl<T: H5Type> DatasetBuilder<T>

pub fn new(parent: &Group) -> Self

Create a new dataset builder and bind it to the parent container.

pub fn packed(&mut self, packed: bool) -> &mut Self

pub fn fill_value(&mut self, fill_value: T) -> &mut Self

pub fn no_chunk(&mut self) -> &mut Self

Disable chunking.

pub fn chunk_auto(&mut self) -> &mut Self

Enable automatic chunking only if chunking is required (default option).

pub fn chunk_infer(&mut self) -> &mut Self

Enable chunking with automatic chunk shape.

pub fn chunk<D: Dimension>(&mut self, chunk: D) -> &mut Self

Set chunk shape manually.

pub fn filters(&mut self, filters: &Filters) -> &mut Self

Set the filters.

pub fn track_times(&mut self, track_times: bool) -> &mut Self

Enable or disable tracking object modification time (disabled by default).

pub fn resizable(&mut self, resizable: bool) -> &mut Self

Make the dataset resizable along all axes (requires chunking).

pub fn gzip(&mut self, level: u8) -> &mut Self

Enable gzip compression with a specified level (0-9).

pub fn szip(&mut self, nn: bool, level: u8) -> &mut Self

Enable szip compression with a specified method (EC, NN) and level (0-32).

If nn if set to true (default), the nearest neighbor method is used, otherwise the method is set to entropy coding.

pub fn shuffle(&mut self, shuffle: bool) -> &mut Self

Enable or disable shuffle filter.

pub fn fletcher32(&mut self, fletcher32: bool) -> &mut Self

Enable or disable fletcher32 filter.

pub fn scale_offset(&mut self, scale_offset: u32) -> &mut Self

Enable scale-offset filter with a specified factor (0 means automatic).

pub fn create<D: Dimension>(&self, name: &str, shape: D) -> Result<Dataset>

Create the dataset and link it into the file structure.

pub fn create_anon<D: Dimension>(&self, shape: D) -> Result<Dataset>

Create an anonymous dataset without linking it.