Module hdf5::file

Structs

Alignment

CacheImageConfig

CacheLogOptions

ChunkCache

CoreDriver

FamilyDriver

File
HDF5 file object.

FileAccess
File access properties.

FileAccessBuilder
Builder used to create file access property list.

FileBuilder
File builder allowing to customize file access/creation property lists.

FileCreate
File creation properties.

FileCreateBuilder
Builder used to create file creation property list.

LibVerBounds

LogFlags

LogOptions

MetadataCacheConfig

MultiDriver

MultiFile

MultiLayout

PageBufferSize

PhaseChangeInfo
Threshold values for storage of shared message indexes.

SharedMessageIndex
Configuration settings for a shared message index.

SharedMessageType
Types of messages that can be stored in a shared message index.

SizeofInfo
Size of the offsets and lengths used in a file.

SplitDriver

SymbolTableInfo
Size of parameters used to control the symbol table nodes.

Enums

CacheDecreaseMode

CacheIncreaseMode

FileCloseDegree

FileDriver

FileSpaceStrategy
File space handling strategy.

FlashIncreaseMode

LibraryVersion

MetadataWriteStrategy

OpenMode
File opening mode.