Struct hdf5::file::FileCreateBuilder

pub struct FileCreateBuilder { /* fields omitted */ }

Builder used to create file creation property list.

Methods

impl FileCreateBuilder

pub fn new() -> Self

Creates a new file creation property list builder.

pub fn from_plist(plist: &FileCreate) -> Result<Self>

Creates a new builder from an existing property list.

pub fn userblock(&mut self, size: u64) -> &mut Self

Sets user block size.

Sets the user block size of a file creation property list. The default user block size is 0; it may be set to any power of 2 equal to 512 or greater (512, 1024, 2048, etc.).

pub fn sym_k(&mut self, tree_rank: u32, node_size: u32) -> &mut Self

Sets the size of parameters used to control the symbol table nodes.

Passing in a value of zero (0) for one of the parameters (tree_rank or node_size) retains the current value.

For further details, see SymbolTableInfo.

pub fn istore_k(&mut self, ik: u32) -> &mut Self

Sets the size of the parameter used to control the B-trees for indexing chunked datasets.

istore_k is one half the rank of a tree that stores chunked raw data. On average, such a tree will be 75% full, or have an average rank of 1.5 times the value of istore_k.

The HDF5 library uses istore_k * 2 as the maximum number of entries before splitting a B-tree node. Since only 2 bytes are used in storing the number of entries for a B-tree node in an HDF5 file, istore_k * 2 cannot exceed 65536.

The default value for istore_k is 32.

pub fn shared_mesg_phase_change(
&mut self,
max_list: u32,
min_btree: u32
) -> &mut Self

Sets shared object header message storage phase change thresholds.

For further details, see PhaseChangeInfo.

pub fn shared_mesg_indexes(
&mut self,
indexes: &[SharedMessageIndex]
) -> &mut Self

Configures shared object header message indexes.

For each specified index, sets the types of messages that may be stored and the minimum size of each message

pub fn file_space_page_size(&mut self, fsp_size: u64) -> &mut Self

Sets the file space page size.

The minimum size is 512. Setting a value less than 512 will result in an error. The library default size for the file space page size when not set is 4096.

pub fn file_space_strategy(&mut self, strategy: FileSpaceStrategy) -> &mut Self

Sets the file space handling strategy and persisting free-space values.

This setting cannot be changed for the life of the file.

For further details, see FileSpaceStrategy.

pub fn finish(&self) -> Result<FileCreate>