Module hdf5::plist::file_create

Structs

FileCreate
File creation properties.

FileCreateBuilder
Builder used to create file creation property list.

PhaseChangeInfo
Threshold values for storage of shared message indexes.

SharedMessageIndex
Configuration settings for a shared message index.

SharedMessageType
Types of messages that can be stored in a shared message index.

SizeofInfo
Size of the offsets and lengths used in a file.

SymbolTableInfo
Size of parameters used to control the symbol table nodes.

Enums

FileSpaceStrategy
File space handling strategy.