Struct hdf5::Dataset

#[repr(transparent)]
pub struct Dataset(_);

Represents the HDF5 dataset object.

Methods

impl Dataset

pub fn is_resizable(&self) -> bool

Returns whether this dataset is resizable along some axis.

pub fn is_chunked(&self) -> bool

Returns whether this dataset has a chunked layout.

pub fn chunks(&self) -> Option<Vec<Ix>>

Returns the chunk shape if the dataset is chunked.

pub fn filters(&self) -> Filters

Returns the filters used to create the dataset.

pub fn tracks_times(&self) -> bool

Returns true if object modification time is tracked by the dataset.

pub fn offset(&self) -> Option<u64>

Returns the absolute byte offset of the dataset in the file if such offset is defined (which is not the case for datasets that are chunked, compact or not allocated yet).

pub fn fill_value<T: H5Type>(&self) -> Result<Option<T>>

Returns default fill value for the dataset if such value is set. Note that conversion to the requested type is done by HDF5 which may result in loss of precision for floating-point values if the datatype differs from the datatype of of the dataset.

pub fn resize<D: Dimension>(&self, d: D) -> Result<()>