Struct hdf5::File

#[repr(transparent)]

pub struct File(_);

Represents the HDF5 file object.

Methods

impl File

pub fn open<P: AsRef<Path>>(filename: P, mode: &str) -> Result<Self>

Create a new file object.

mode 	 File access mode

r 	 Read-only, file must exist

r+ 	 Read/write, file must exist

w 	 Create file, truncate if exists

w-, x 	 Create file, fail if exists

a 	 Read/write if exists, create otherwise

pub fn with_options() -> FileBuilder

pub fn size(&self) -> u64

Returns the file size in bytes (or 0 if the file handle is invalid).

pub fn free_space(&self) -> u64

Returns the free space in the file in bytes (or 0 if the file handle is invalid).

pub fn is_read_only(&self) -> bool

Returns true if the file was opened in a read-only mode.

pub fn userblock(&self) -> u64

Returns the userblock size in bytes (or 0 if the file handle is invalid).

pub fn flush(&self) -> Result<()>

Flushes the file to the storage medium.

pub fn close(self)

Closes the file and invalidates all open handles for contained objects.

pub fn get_create_plist(&self) -> Result<FileCreate>